refactor: board out of game
