+++ src/pieces/knight.rs
@@ -1,7 +1,7 @@
 use super::{Movable, PieceColor, PieceMove, Position};
 use crate::constants::DisplayMode;
+use crate::game::board::GameBoard;
 use crate::game::coord::Coord;
-use crate::game::game::GameBoard;
 use crate::utils::{cleaned_positions, impossible_positions_king_checked, is_cell_color_ally};
 pub struct Knight;
